Deprecate and alias ClassMetadataInfo (#1733)
