+++ lib/Doctrine/ODM/MongoDB/Aggregation/Stage/GraphLookup.php
@@ -24,7 +24,6 @@ use Doctrine\MongoDB\Aggregation\Stage as BaseStage;
 use Doctrine\ODM\MongoDB\Aggregation\Builder;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
 use Doctrine\ODM\MongoDB\Mapping\MappingException;
 use Doctrine\ODM\MongoDB\Types\Type;
 
@@ -198,9 +197,9 @@ class GraphLookup extends BaseStage\GraphLookup
         }
 
         switch ($mapping['storeAs']) {
-            case ClassMetadataInfo::REFERENCE_STORE_AS_ID:
-            case ClassMetadataInfo::REFERENCE_STORE_AS_REF:
-                return ClassMetadataInfo::getReferenceFieldName($mapping['storeAs'], $mapping['name']);
+            case ClassMetadata::REFERENCE_STORE_AS_ID:
+            case ClassMetadata::REFERENCE_STORE_AS_REF:
+                return ClassMetadata::getReferenceFieldName($mapping['storeAs'], $mapping['name']);
                 break;
 
             default:
