+++ lib/Doctrine/ODM/MongoDB/Mapping/Annotations/ReferenceMany.php
@@ -19,7 +19,7 @@
 
 namespace Doctrine\ODM\MongoDB\Mapping\Annotations;
 
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Doctrine\ODM\MongoDB\Utility\CollectionHelper;
 
 /**
@@ -32,7 +32,7 @@ final class ReferenceMany extends AbstractField
     public $type = 'many';
     public $reference = true;
     public $simple = false; // @deprecated
-    public $storeAs = ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF_WITH_DB;
+    public $storeAs = ClassMetadata::REFERENCE_STORE_AS_DB_REF_WITH_DB;
     public $targetDocument;
     public $discriminatorField;
     public $discriminatorMap;
