+++ lib/Doctrine/ODM/MongoDB/Proxy/ProxyFactory.php
@@ -28,6 +28,7 @@ use Doctrine\Common\Proxy\ProxyGenerator;
 use Doctrine\Common\Util\ClassUtils;
 use Doctrine\Common\Proxy\Proxy as BaseProxy;
 use Doctrine\Common\Persistence\Mapping\ClassMetadata as BaseClassMetadata;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Doctrine\ODM\MongoDB\Persisters\DocumentPersister;
 use Doctrine\ODM\MongoDB\Utility\LifecycleEventManager;
 use ReflectionProperty;
@@ -87,7 +88,7 @@ class ProxyFactory extends AbstractProxyFactory
      */
     public function skipClass(BaseClassMetadata $class)
     {
-        /* @var $class \Doctrine\ODM\Mongodb\Mapping\ClassMetadataInfo */
+        /* @var $class ClassMetadata */
         return $class->isMappedSuperclass || $class->isQueryResultDocument || $class->getReflectionClass()->isAbstract();
     }
 
@@ -96,7 +97,7 @@ class ProxyFactory extends AbstractProxyFactory
      */
     public function createProxyDefinition($className)
     {
-        /* @var $classMetadata \Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo */
+        /* @var $classMetadata ClassMetadata */
         $classMetadata     = $this->metadataFactory->getMetadataFor($className);
         $documentPersister = $this->uow->getDocumentPersister($className);
         $reflectionId      = $classMetadata->reflFields[$classMetadata->identifier];
