+++ lib/Doctrine/ODM/MongoDB/Query/Builder.php
@@ -21,7 +21,7 @@ namespace Doctrine\ODM\MongoDB\Query;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Hydrator;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 
 /**
  * Query builder for ODM.
@@ -407,7 +407,7 @@ class Builder extends \Doctrine\MongoDB\Query\Builder
             $query['distinct'] = $documentPersister->prepareFieldName($query['distinct']);
         }
 
-        if ($this->class->inheritanceType === ClassMetadataInfo::INHERITANCE_TYPE_SINGLE_COLLECTION && ! empty($query['upsert']) &&
+        if ($this->class->inheritanceType === ClassMetadata::INHERITANCE_TYPE_SINGLE_COLLECTION && ! empty($query['upsert']) &&
             (empty($query['query'][$this->class->discriminatorField]) || is_array($query['query'][$this->class->discriminatorField]))) {
             throw new \InvalidArgumentException('Upsert query that is to be performed on discriminated document does not have single ' .
                 'discriminator. Either not use base class or set \'' . $this->class->discriminatorField . '\' field manually.');
@@ -415,7 +415,7 @@ class Builder extends \Doctrine\MongoDB\Query\Builder
 
         if ( ! empty($query['select'])) {
             $query['select'] = $documentPersister->prepareSortOrProjection($query['select']);
-            if ($this->hydrate && $this->class->inheritanceType === ClassMetadataInfo::INHERITANCE_TYPE_SINGLE_COLLECTION
+            if ($this->hydrate && $this->class->inheritanceType === ClassMetadata::INHERITANCE_TYPE_SINGLE_COLLECTION
                 && ! isset($query['select'][$this->class->discriminatorField])) {
                 $includeMode = 0 < count(array_filter($query['select'], function($mode) { return $mode == 1; }));
                 if ($includeMode && ! isset($query['select'][$this->class->discriminatorField])) {
