+++ lib/Doctrine/ODM/MongoDB/Query/Expr.php
@@ -21,7 +21,6 @@ namespace Doctrine\ODM\MongoDB\Query;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
 use Doctrine\ODM\MongoDB\Mapping\MappingException;
 
 /**
@@ -77,20 +76,20 @@ class Expr extends \Doctrine\MongoDB\Query\Expr
             $storeAs = array_key_exists('storeAs', $mapping) ? $mapping['storeAs'] : null;
 
             switch ($storeAs) {
-                case ClassMetadataInfo::REFERENCE_STORE_AS_ID:
+                case ClassMetadata::REFERENCE_STORE_AS_ID:
                     $this->query[$mapping['name']] = $reference;
                     return $this;
                     break;
 
-                case ClassMetadataInfo::REFERENCE_STORE_AS_REF:
+                case ClassMetadata::REFERENCE_STORE_AS_REF:
                     $keys = ['id' => true];
                     break;
 
-                case ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF:
-                case ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF_WITH_DB:
+                case ClassMetadata::REFERENCE_STORE_AS_DB_REF:
+                case ClassMetadata::REFERENCE_STORE_AS_DB_REF_WITH_DB:
                     $keys = ['$ref' => true, '$id' => true, '$db' => true];
 
-                    if ($storeAs === ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF) {
+                    if ($storeAs === ClassMetadata::REFERENCE_STORE_AS_DB_REF) {
                         unset($keys['$db']);
                     }
 
@@ -129,20 +128,20 @@ class Expr extends \Doctrine\MongoDB\Query\Expr
             $storeAs = array_key_exists('storeAs', $mapping) ? $mapping['storeAs'] : null;
 
             switch ($storeAs) {
-                case ClassMetadataInfo::REFERENCE_STORE_AS_ID:
+                case ClassMetadata::REFERENCE_STORE_AS_ID:
                     $this->query[$mapping['name']] = $reference;
                     return $this;
                     break;
 
-                case ClassMetadataInfo::REFERENCE_STORE_AS_REF:
+                case ClassMetadata::REFERENCE_STORE_AS_REF:
                     $keys = ['id' => true];
                     break;
 
-                case ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF:
-                case ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF_WITH_DB:
+                case ClassMetadata::REFERENCE_STORE_AS_DB_REF:
+                case ClassMetadata::REFERENCE_STORE_AS_DB_REF_WITH_DB:
                     $keys = ['$ref' => true, '$id' => true, '$db' => true];
 
-                    if ($storeAs === ClassMetadataInfo::REFERENCE_STORE_AS_DB_REF) {
+                    if ($storeAs === ClassMetadata::REFERENCE_STORE_AS_DB_REF) {
                         unset($keys['$db']);
                     }
