+++ lib/Doctrine/ODM/MongoDB/Utility/CollectionHelper.php
@@ -18,7 +18,7 @@
  */
 
 namespace Doctrine\ODM\MongoDB\Utility;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 
 /**
  * Utility class used to unify checks on how collection strategies should behave.
@@ -28,7 +28,7 @@ use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
  */
 class CollectionHelper
 {
-    const DEFAULT_STRATEGY = ClassMetadataInfo::STORAGE_STRATEGY_PUSH_ALL;
+    const DEFAULT_STRATEGY = ClassMetadata::STORAGE_STRATEGY_PUSH_ALL;
 
     /**
      * Returns whether update query must be included in query updating owning document.
@@ -38,7 +38,7 @@ class CollectionHelper
      */
     public static function isAtomic($strategy)
     {
-        return $strategy === ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET || $strategy === ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET_ARRAY;
+        return $strategy === ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET || $strategy === ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET_ARRAY;
     }
 
     /**
@@ -49,7 +49,7 @@ class CollectionHelper
      */
     public static function isHash($strategy)
     {
-        return $strategy === ClassMetadataInfo::STORAGE_STRATEGY_SET || $strategy === ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET;
+        return $strategy === ClassMetadata::STORAGE_STRATEGY_SET || $strategy === ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET;
     }
     
     /**
@@ -60,7 +60,7 @@ class CollectionHelper
      */
     public static function isList($strategy)
     {
-        return $strategy !== ClassMetadataInfo::STORAGE_STRATEGY_SET && $strategy !== ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET;
+        return $strategy !== ClassMetadata::STORAGE_STRATEGY_SET && $strategy !== ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET;
     }
     
     /**
@@ -74,10 +74,10 @@ class CollectionHelper
         return in_array(
             $strategy,
             [
-                ClassMetadataInfo::STORAGE_STRATEGY_SET,
-                ClassMetadataInfo::STORAGE_STRATEGY_SET_ARRAY,
-                ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET,
-                ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET_ARRAY
+                ClassMetadata::STORAGE_STRATEGY_SET,
+                ClassMetadata::STORAGE_STRATEGY_SET_ARRAY,
+                ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET,
+                ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET_ARRAY
             ]
         );
     }
