+++ tests/Doctrine/ODM/MongoDB/Tests/Functional/CustomCollectionsTest.php
@@ -5,7 +5,7 @@ namespace Doctrine\ODM\MongoDB\Tests\Functional;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ODM\MongoDB\Configuration;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Doctrine\ODM\MongoDB\PersistentCollection;
 use Doctrine\ODM\MongoDB\PersistentCollection\PersistentCollectionInterface;
 use Documents\File;
@@ -33,7 +33,7 @@ class CustomCollectionsTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
      */
     public function testCollectionClassHasToImplementCommonInterface()
     {
-        $cm = new ClassMetadataInfo('stdClass');
+        $cm = new ClassMetadata('stdClass');
 
         $cm->mapField(array(
             'fieldName' => 'assoc',
