+++ tests/Doctrine/ODM/MongoDB/Tests/Functional/NestedCollectionsTest.php
@@ -3,7 +3,7 @@
 namespace Doctrine\ODM\MongoDB\Tests\Functional;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadataInfo;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Documents\Phonebook;
 use Documents\Phonenumber;
 
@@ -64,12 +64,12 @@ class NestedCollectionsTest extends \Doctrine\ODM\MongoDB\Tests\BaseTest
     public function provideStrategy()
     {
         return array(
-            array(ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET),
-            array(ClassMetadataInfo::STORAGE_STRATEGY_ATOMIC_SET_ARRAY),
-            array(ClassMetadataInfo::STORAGE_STRATEGY_SET),
-            array(ClassMetadataInfo::STORAGE_STRATEGY_SET_ARRAY),
-            array(ClassMetadataInfo::STORAGE_STRATEGY_PUSH_ALL),
-            array(ClassMetadataInfo::STORAGE_STRATEGY_ADD_TO_SET),
+            array(ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET),
+            array(ClassMetadata::STORAGE_STRATEGY_ATOMIC_SET_ARRAY),
+            array(ClassMetadata::STORAGE_STRATEGY_SET),
+            array(ClassMetadata::STORAGE_STRATEGY_SET_ARRAY),
+            array(ClassMetadata::STORAGE_STRATEGY_PUSH_ALL),
+            array(ClassMetadata::STORAGE_STRATEGY_ADD_TO_SET),
         );
     }
 }
